Wifi-Hal: Fix NanCapabilities status field
The status field is modified to u32 from u16. Fix the same in
HAL.

BUG=26918148

+++ qcwcn/wifi_hal/nan_i.h
@@ -950,7 +950,7 @@ typedef struct PACKED
 {
     NanMsgHeader fwHeader;
     /* status of the request */
-    u16 status;
+    u32 status;
     u32 value;
     u32 max_concurrent_nan_clusters;
     u32 max_publishes;
